Fix warnings (incl. undefined behavior) and some optimization  * pass 'make CPPFLAGS="-Wall -Werror -DGYOTO_NO_DEPRECATED \                         -Wno-error=literal-suffix -Wno-error=unknown-pragmas"'  * some code optimization in NumericalMetricLorene

+++ yorick/gyoto_SpectroComplex.C
@@ -64,7 +64,7 @@ void YGyoto::SpCplxEval(SmartPointer<Spectrometer::Generic> *OBJ_, int argc) {
     if ((*rvset)++) y_error(rmsg);
     ++(*paUsed);
     size_t cardinal=(*OBJ) -> getCardinal();
-    long item = ygets_l(iarg);
+    size_t item = ygets_l(iarg);
     if (item > cardinal) y_error("index overreach array bounds");
     if (item <= 0) {
       item += cardinal;
